node: move wormchain watcher
